Refactor GPU painter

+++ src/brush_settings.h
@@ -2,12 +2,11 @@
 
 #include "cuda_runtime.h"
 
-#include "color.h"
 
 struct BrushSettings {
-	double pressure;
-	double heightPressure;
-	double size;
-	double falloff;
+	float pressure;
+	float heightPressure;
+	float size;
+	float falloff;
 	float3 color;
 };
